test: 테스트 fixture 분리
